🔨 Simplify custom variant build

+++ buildroot/share/PlatformIO/scripts/STM32F1_create_variant.py
@@ -3,28 +3,40 @@
 #
 import pioutil
 if pioutil.is_pio_build():
-    import shutil, marlin
     from pathlib import Path
-    env = pioutil.env
-    platform = env.PioPlatform()
-    board = env.BoardConfig()
 
-    FRAMEWORK_DIR = Path(platform.get_package_dir("framework-arduinoststm32-maple"))
-    assert FRAMEWORK_DIR.is_dir()
-
-    source_root = Path("buildroot/share/PlatformIO/variants")
+    source_root_str = "buildroot/share/PlatformIO/variants"
+    source_root = Path(source_root_str)
     assert source_root.is_dir()
 
+    env = pioutil.env
+    board = env.BoardConfig()
     variant = board.get("build.variant")
-    variant_dir = FRAMEWORK_DIR / "STM32F1/variants" / variant
 
     source_dir = source_root / variant
     assert source_dir.is_dir()
 
-    if variant_dir.is_dir():
-        shutil.rmtree(variant_dir)
+    if True:
 
-    if not variant_dir.is_dir():
-        variant_dir.mkdir()
+        platform = env.PioPlatform()
+        FRAMEWORK_DIR = Path(platform.get_package_dir("framework-arduinoststm32-maple"))
+        assert FRAMEWORK_DIR.is_dir()
 
-    marlin.copytree(source_dir, variant_dir)
+        variant_dir = FRAMEWORK_DIR / "STM32F1/variants" / variant
+
+        if variant_dir.is_dir():
+            import shutil
+            shutil.rmtree(variant_dir)
+
+        if not variant_dir.is_dir():
+            variant_dir.mkdir()
+
+        import marlin
+        marlin.copytree(source_dir, variant_dir)
+
+    else:
+
+        # The following almost works, but __start__ (from wirish/start.S) is not seen by common.inc
+        board.update("build.variants_dir", source_root_str);
+        src = str(source_dir)
+        env.Append(BUILD_FLAGS=[f"-I{src}", f"-L{src}/ld"])  # Add include path for variant
